chore(deps): do not enable default features of chrono
Enabling the default features of chrono causes the time 0.1 crate to be
added as a transitive dependency.

This old version of time is affected by CVE RUSTSEC-2020-0071

Thanks to work done inside of chrono 0.4, there are high chances the
majority of the codebases do not actually need it. Which is exactly
the case for picky.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>