ci: aggregate success results into a single job (#288)
This greatly simplifies the terraform configuration, because we only
have to require the "Success" status check. It’s simpler than listing
all the checks in terraform which can be sometimes problematic:

- We need to list all the job names generated using `matrix`
- We need to open a PR in infrastructure-as-code AND to deploy the
  change when we change / add / remove a job

Here, we just list the required jobs directly as dependencies of the
`success` job, without having to modify multiple repositories. Job
dependencies also does not use the job names, but the job "key", so
there is much less things to enumerate. Less error prone overall.