Change request target handling to comply with draft-cavage-http-signatures-12 when signing HTTP methods.
For reasons unclear to me, the `http-signature` module treats the request
target specially in that it is added to & read from the signing string with
no header name (unlike, say, `(created)` or `(expires)`). As we see
[here](https://datatracker.ietf.org/doc/html/draft-cavage-http-signatures-12#page-9),
it, too,  has a "psuedo-header" name: `(request-target)`.

It's absence from the signing string was breaking my attempts to
use picky to validate message signatures from Mastodon.